[CI/Build]: Add project metadata to pyproject toml file (#592)
* Add project metadata to pyproject toml

The pyproject.toml file can handle most the project metadata
in a more central location to the `setup.py` file. This commit
moves it to the toml file.

It also adds dynamic versioning which uses git tags.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Update files after ruff check

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Fix pyproject toml file after rebase

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Add setuptools to requirements

This is neded by the build system as it may use any version
of setuptools whihc may not be compatible.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Update PR template to remove old tagging information

Now using dynamic versioning, no need anymore to add tags.
This only needs to be done now at release time.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Update after review

Review comments:
- https://github.com/LMCache/LMCache/pull/592#pullrequestreview-2823447030

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

---------

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>
Signed-off-by: zhuohangu <zhuohan@uchicago.edu>