[CI/Build] Add publishing to TestPyPi and refactor PyPi publish (#611)
* Add publishing to TestPyPi and refactor PyPi publish

Added publishing to TestPyPi on every PR merge and also
a release. This will give regular intermediatery builds between
releases.

In addition did some refactoring for improvement and security reasons
as follows:

- Separated build from the publish jobs
- Added security action to the runner
- Specified latest action versions to use
- Replaced GH release action with the `gh` CLI command directly
- Replaced `twine` command with pypi publish action

Note: The publishing no longer requires tokens for publishing to PyPi and TestPyPi.
It insteads uses "Trusted Publishers": https://docs.pypi.org/trusted-publishers/creating-a-project-through-oidc/
This reduces the steps for automated publishing in the workflow.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Add text to describe process for publishing

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

---------

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>
Signed-off-by: zhuohangu <zhuohan@uchicago.edu>