Micro benchmark for VLLMPagedMemGPUConnectorV2.toGpu and a performance fix (#678)
* gpu_connector: add VLLMPagedConnectorV2.to_gpu micro-benchmark

We've noticed it spends a lot of time in
VLLMPagedConnectorV2._pointers_are_good, so before trying to improve it,
adding a micro-benchmark.

* pagedMemGPUConnectorV2: optimize to/from GPU

There was a number of issues with `_pointersAreGood` method, causing
it to massively dominate the to/from GPU cost.

First, why even check it, if the check takes linear time at worst, and
the initialization is also linear?

Then, getting/setting elements of a tensor is suboptimal, let's do it
in a single assignment instead.

And finally, the actual multi-layer transfer kernel runs on GPU, so
let's move pointers there.

This gives 3.3x speedup in the microbenchmark, see the results here:
https://gist.github.com/yanok/685182fe15b46bf0d99797853a4f9f9a

I only changed the VLLMPagedMemGPUConnectorV2 class, since that's the
one we are currently using in our setup. But I saw similar code in other
GPU connector classes, we might want to patch them as well.

Signed-off-by: zhuohangu <zhuohan@uchicago.edu>