Remove duplicate requirements loading in setup tools
RProject requirements are loaded from pyproject.toml since PR #728.
However, #702 was not rebased with changes and therefore
re-introduced the loading of project requirements again in setup tools.

This results in a warning message when install the project as follows:
/tmp/pip-build-env-3fib4lf4/overlay/lib/python3.12/site-packages/setuptools/config/_apply_pyprojecttoml.py:82:
SetuptoolsWarning: `install_requires` overwritten in `pyproject.toml` (dependencies)

This commit rollback the unnecessary duplication.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>