Add SPDX license headers
A Python script is added which checks Python files for the license header
and add it if it is missing, it will add the header. This script is added as a local
pre-commit hook.

The Python script is copied from vLLM repository on GitHub.
For the original code, see [vLLM](https://github.com/vllm-project/vllm/blob/main/tools/check_spdx_header.py)

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>