[Core] fix nixl_storage_backend assert
an assert is hit when local_cpu_backend tries to allocate, because
allocator type is not MixedMemoryAllocator.

In order to adapt to this assert, change nixl_storage_backend
to use MixedMemoryAllocator with paging=true, which will create a
PagedTensorMemoryAllocator internally. then use this internal allocator
for NIXL instead of creating PagedTensorMemoryAllocator.

Signed-off-by: Tomer Shmilovich <tshmilovich@nvidia.com>