Remove Python 3.9 as the minimum Python version
LMCache was built on Python 3.10 and as such has keywords like
match which are not compatible in Python 3.9. We therefore
would need to make changes to the code base. Python 3.9
will no longer be supported next month. Therefore, it is more
practical to support 3.10 as the minimum version.

This is a revert of some of the changes in:
https://github.com/LMCache/LMCache/pull/1440

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>