fs_connector: support two-tier fanout tree to balance the kv cache files
As the KV Cache files grow, the existing directory structure is prone to
generating extremely large directories, which poses significant pressure
on file lookup and traversal. Therefore, we adjust the single directory
tree structure to a two-tier hash directory tree structure, allowing for
a more even distribution of KV Cache files and avoiding the creation of
large directories.

Signed-off-by: Hongbo Li <lihb2113@outlook.com>